+++ spiel/demo/demo.py
@@ -17,8 +17,8 @@ from rich.syntax import Syntax
 from rich.text import Text
 
 from spiel import __version__
-from spiel.app import Deck, Slide
-from spiel.image import Image
+from spiel.deck import Deck, Slide
+from spiel.renderables.image import Image
 
 deck = Deck(name=f"Spiel Demo Deck (v{__version__})")
