+++ spiel/exceptions.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+
+class SpielException(Exception):
+    pass
+
+
+class NoDeckFound(SpielException):
+    pass
