+++ spiel/widgets/version.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from rich.console import RenderableType
+from textual.widget import Widget
+
+
+class DebugInfo(Widget):
+    DEFAULT_CSS = """
+    VersionPanel {
+        width: auto;
+        height: auto;
+    }
+    """
+
+    def render(self) -> RenderableType:
+        return
