+++ tests/test_image.py
@@ -2,8 +2,8 @@ import pytest
 from PIL import Image as Img
 from rich.console import Console
 
-from spiel.image import Image, ImageSize
 from spiel.main import DEMO_DIR
+from spiel.renderables.image import Image, ImageSize
 
 
 @pytest.fixture
